requirements.txt
langgraph>=0.0.19
openai>=1.3.0
flask>=2.0.0
python-dotenv>=1.0.0
duckduckgo-search>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.28.0
langfuse>=3.0.0
